fix sdk tests
