+++ apps/sdk/tests/test_db_connector.py
@@ -1,27 +1,10 @@
 import pytest
 from unittest.mock import MagicMock, patch
-from sdk.db_connector import DBConnector
+from db_connector import DBConnector
 
 """
 This module contains the tests for the DBConnector.
 """
-
-
-
-@pytest.fixture(scope="function")
-def mock_db_connector():
-    """
-    Fixture to create a mock DBConnector instance.
-    """
-    with patch("sdk.db_connector.psycopg2.connect") as mock_connect:
-        mock_conn = MagicMock()
-        mock_cursor = MagicMock()
-        mock_connect.return_value = mock_conn
-        mock_conn.cursor.return_value = mock_cursor
-        db_connector = DBConnector()
-        yield db_connector
-        db_connector.close_connection()
-
 
 def test_connect_to_db(mock_db_connector):
     """
@@ -85,12 +68,20 @@ def test_get_loan_state(mock_db_connector):
     """
     Test the get_loan_state method.
     """
-    mock_db_connector.cur.fetchone.return_value = ["loan_state"]
+    mock_db_connector.cur.fetchone.return_value = [
+        {"collateral": 100.0},  # collateral
+        {"debt": 50.0},         # debt
+        {"deposit": 200.0}      # deposit
+    ]
     result = mock_db_connector.get_loan_state("protocol_id", "wallet_id")
-    assert result == "loan_state"
+    assert result == {
+        "collateral": {"collateral": 100.0},
+        "debt": {"debt": 50.0},
+        "deposit": {"deposit": 200.0}
+    }
     mock_db_connector.cur.execute.assert_called_once_with(
         """
-                SELECT * FROM loan_state
+                SELECT collateral, debt, deposit FROM loan_state
                 WHERE protocol_id = %s and "user" = %s;
             """,
         ("protocol_id", "wallet_id"),
@@ -103,7 +94,7 @@ def test_get_loan_state_none(mock_db_connector):
     """
     mock_db_connector.cur.fetchone.return_value = None
     result = mock_db_connector.get_loan_state("protocol_id", "wallet_id")
-    assert result is None
+    assert result == None
 
 
 def test_close_connection(mock_db_connector):
